FogBugz #168803: support to configure initial state of bridges
This change adds support to the Altera fpga2sdram and hps2fgpa bridge
drivers for a new optional device tree property, init-val. This property
configures the driver to enable or disable the bridge when instantiated.
If the property does not exist, the driver will leave the bridge in
its current state.

Signed-off-by: Matthew Gerlach <mgerlach@altera.com>