FogBugz #198256: Fix unnecessary USB overcurrent condition
The CycloneV and ArriaV devkit has a small C_VBUS capacitor on the devkit.
Because of this, the USB driver sometimes detects an unnecessary
overcurrent error condition.

This patch uses the external VbusValid signal instead.

Updated for v4.5 so that a core reset will get called to prevent an
overcurrent condition on the Arria10.

Signed-off-by: Dinh Nguyen <dinguyen@altera.com>