FogBugz #293560-1: A10SYCON: Cleanup and additions to hwmon
In 3.10LTSI, this was FogBugz#27094-17 but this seems to have
gotten lost in the migration. Pull it in here.

Cleanup of the bitfields. Remove labels for resets and enables.
Addition of HPS resets and peripheral reset bitfields.
The MAX5 hwmon was originally implemented with only bitfield reads.
It is useful to add the complete register read of the MAX5 registers
to this module.

Signed-off-by: Thor Thayer <tthayer@opensource.altera.com>