FogBugz #282531-1: Move the location of setting the SD PWREN bit
We were writing to the sdmmc register that toggles the PWREN bit when the
clock to the IP could be turned off, thus would result in a hang.

This patch moves the location of turning on PWREN to the set_ios function,
which is guaranteed to have the clock turned on. The new location of where
toggling the PWREN bit takes place in the same function of the driver
(dw_mci_set_ios), but just does it in the SOCFPGA platform specific
portion of the driver.

The "pwr-en" property can be used to determine if the SD power enable line
should be turned on. In order to support the Arria10 Rev. A devkit, we
should be turning on/off PWREN that is cued off of "pwr-en".

This patch will assume that "pwr-en" is 1 when the property is not
populated, which will be the case for Cyclone5/Arria5 based devkits. For
the Arria10 Revision A devkit, which has a reverse PWREN logic, so we want
to leave the PWREN bit to 0, thus we need "pwr-en = <0>".

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
---
v3: reworked if statement around pwr_en
v2: update commit message, update license header and change a dev_info to
    dev_dbg