FogBugz #392349-5: ARM64: dts: Stratix10 SWVP device tree release version.
This is the first deliverable version of the S10 SWVP dts.

This is one of the patches for Stratix10 SWVP.  Normally it would be
applied to the latest kernel version, but v4.6 accesses an ARM64 register
that doesn't exist in the QEMU/Mentor model, and thereby causes a kernel
panic.  So this patch is being applied to v4.5 in order to meet the SWVP
deadline.  Once the QEMu/Mentor model is fixed, then these patches will
be migrated to the latest kernel.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>
---
V2: Use Intel copyright instead of Altera