FogBugz #392349-6: ARM64: dts: Remove usbphy from Stratix10 SWVP dts.
The generic phy framework doesn't need a dts entry.  In fact, when
CONFIG_GENERIC_PHY is defined, the DWC2 USB driver won't load if there is
a usbphy in the device tree.

This is one of the patches for Stratix10 SWVP.  Normally it would be
applied to the latest kernel version, but v4.6 accesses an ARM64 register
that doesn't exist in the QEMU/Mentor model, and thereby causes a kernel
panic.  So this patch is being applied to v4.5 in order to meet the SWVP
deadline.  Once the QEMu/Mentor model is fixed, then these patches will
be migrated to the latest kernel.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>