FogBugz #392349-7: ARM64: dts: Enable watchdog in Stratix10 SWVP
Enable one watchdog timer in the Stratix10 SWVP device tree
by adding a clocks entry and setting status to okay.  Note that
the pre-v4.6 Designware Watchdog driver only supports one watchdog
timer.

This is one of the patches for Stratix10 SWVP.  Normally it would be
applied to the latest kernel version, but v4.6 accesses an ARM64 register
that doesn't exist in the QEMU/Mentor model, and thereby causes a kernel
panic.  So this patch is being applied to v4.5 in order to meet the SWVP
deadline.  Once the QEMu/Mentor model is fixed, then these patches will
be migrated to the latest kernel.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>