ARM: dts: gose: Add da9063 PMIC device node for system restart
Enable i2c6, and add a device node for the da9063 PMIC, with subnodes
for rtc and wdt.  Regulator support is not yet included.

This allows the system to be restarted when the watchdog timer times
out, or when a system restart is requested.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit c20839a36132462496f939b0d13afee009c9a547)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>