ARM: dts: r8a7745: Link ARM GIC to clock and clock domain
Link the ARM GIC to the INTC-SYS module clock, and add it to the SYSC
"always-on" PM Domain, so it can be power managed using that clock.

Note that currently the GIC-400 driver doesn't support module clocks nor
Runtime PM, so this must be handled as a critical clock.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit db017f399639f68827edc954205803272ef20b24)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>