FogBugz #493306: mark the l4_sp_clk as a critical clock
The l4_sp_clk feeds a 100 MHz clock to various peripherals, such as
SP timers, UARTs, and I2C. Most drivers disable the clock in order
to get the rate, which causes the system to hang. Thus we need to mark
this clock as a CRITICAL clock that cannot get gated.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>