FogBugz #560428: fix up sdmmc_clk's parents
The clock documentation is a bit confusing. Had to resort to Quartus to
get a clear sense of the clocking on Stratix10. The main_sdmmc_clk is the
parent clock for the sdmmc_clk. The main_sdmmc_clk's parent is the
main_noc_base_clk. This patch gets the correct clock parent for the SDMMC.

main_noc_base_clk
	|->main_sdmmc_clk
   		|->sdmmc_free_clk
			|->sdmmc_clk

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>