HSD #1707052006: Add initial support for Agilex devkit
Before the clock driver can be finished on Agilex, we hack up the clock
frequencies for the uart/timer/sdmmc/gmacs, i.e. hard code them the
frequencies.

We add a agilex_minimal_defconfig, that will be deleted in 1 or 2
kernel versions.

We add the reset driver support for Agilex, and add dma-ranges to limit
of 2GB.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
---
FYI: This commit will be removed within 1 or 2 kernel versions as upstream
     full support for the Agilex platform will be completed.
v2: Add frequencies for sdmmc and gmacs