fpga: bridge: incorrect set to clear freeze_illegal_request register
A Partial Region Controller can be connected to one or more
Freeze Bridge. Each Freeze Bridge has an illegal_request
bit represented in the freeze_illegal_request register.
Thus, instead of just set to clear the illegal_request bit
for first Freeze Bridge, we need to ensure the set to clear
action is applied to which ever Freeze Bridge that has
occurrence of illegal request.

Signed-off-by: Chew, Chiau Ee <chiau.ee.chew@intel.com>