HSD#16019035440-1 crypto: intel_fcs: Support new SDOS decrypt error code (#116)
There is a change in SDM mailbox respond for SDOS decryption. SDM have new
error code, 0x102 and 0x103, to notify user the decryption is successful
but with warning. Skip error handling if status is ERROR_CODE_INVALID_KEY
or ERROR_CODE_OUTDATED_KEY to prevent EFAULT return.

Signed-off-by: Teh Wen Ping <wen.ping.teh@intel.com>

State: not-upstreamable