HSD #16020988342: net: stmmac: Fix FPE handshaking flow
In the current FPE handshaking flow, FPE is enabled for local only
the link partner sends Verify mPacket and the local responds with a
Response mPacket. i.e., FPE is only enabled if lo_state of FPE
handshaking flow enters FPE_STATE_ENTERING_ON. But this state is
not relevant for the local.

802.3br mentions the MAC merge sublayer receiver always operates the
same way regardless of whether the preemption is active or not. In
other words, the receiver is always ready for preemption. DWMAC IP
receive side pre-emption is always active. Pre-emption capability is
activated independently on each end. So for enabling transmit side
pre-emption, no need to wait for Verify mPacket from the link partner
Response mPacket from the link partner is enough.

Also seperated FPE receive and transmit configuration as pre-emption
is activated independently on each end. Also pMAC is always ready
for receive pre-emptible traffic.

Signed-off-by: Rohan G Thomas <rohan.g.thomas@intel.com>