Initial commit for linux socfpga CI processes
This is the initial commit for the linux socfpga CI processes.
This commit adds a CODEOWNERS file that add the reviewers automatically
when a PR is submitted.
This commit also adds the pipeline yml file that runs the github action
checks.
This commit also adds a template for the Pull request.

Signed-off-by: Adrian Ng Ho Yin <adrianhoyin.ng@altera.com>